CRUD for Creator, genre and publisher.
Additionally an attempt at the cookies and authentication but I have been banging my head for hours and can't figure it out. FIgured I'd reach out for help. I will keep on looking at it tonight and tomorrow but I think the CRUD is pretty good. Not pretty like Adam would make it but it gets the job done.